+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next version
+
+-   Session settings: show clock while dragging the slider.
+
 # Version 1.8.15
 
 -   Switch to a new service worker imp,ementation that does not trigger